% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloszlasPlots.R
\name{create_discrete_eloszlas_plot}
\alias{create_discrete_eloszlas_plot}
\title{Create Discrete Eloszlas Plot}
\usage{
create_discrete_eloszlas_plot(
  data_to_plot,
  title_to_plot,
  caption_to_plot,
  plot_params = list(x_felirat = "X-axis", fill_felirat = "Fill Legend", fig.height = 6,
    fig.width = 13)
)
}
\description{
Create Discrete Eloszlas Plot
}
