% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloszlasPlots.R
\name{create_plot}
\alias{create_plot}
\title{Create Plot}
\usage{
create_plot(
  plot_type,
  data_to_plot,
  title_to_plot,
  caption_to_plot,
  plot_params = list(x_felirat = "X-axis", fill_felirat = "Fill Legend", fig.height = 6,
    fig.width = 13)
)
}
\description{
Create Plot
}
