% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_objects.R
\name{get_echarts}
\alias{get_echarts}
\title{Create echarts for SurveyQuestion object}
\usage{
get_echarts(question, ...)
}
\arguments{
\item{question}{A SurveyQuestion object}

\item{...}{Additional arguments (unused)}
}
\value{
An echarts object
}
\description{
Create echarts for SurveyQuestion object
}
