% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_objects.R
\name{get_ggplot}
\alias{get_ggplot}
\title{Create ggplot for SurveyQuestion object}
\usage{
get_ggplot(question, ...)
}
\arguments{
\item{question}{A SurveyQuestion object}

\item{...}{Additional arguments (unused)}
}
\value{
A ggplot object
}
\description{
Create ggplot for SurveyQuestion object
}
