% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_latest_table.R
\name{get_latest_table}
\alias{get_latest_table}
\title{Get latest tables from an Oracle scheme}
\usage{
get_latest_table(
  table_name_val,
  sema_name_val,
  tibble_of_tables,
  min_date = NULL,
  max_date = NULL,
  append = ""
)
}
\arguments{
\item{table_name_val}{Character string. Filtering string to table names}

\item{sema_name_val}{Character string. Filtering string to schema names}

\item{tibble_of_tables}{Tibble. If previously all possible names are extracted, they can be set as default.}

\item{min_date}{Date string (YYMMDD), will be converted into date. Gives the minimum date for the search.}

\item{max_date}{Date string (YYMMDD), will be converted into date. Gives the maximum date for the search.}

\item{append}{append a text after the table}
}
\value{
A tibble with given table name patterns and their latest table names in an Oracle scheme.
}
\description{
\code{get_latest_table()} gets latest table lists from Oracle connections.
}
