% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_optimal_font_color.r
\name{get_optimal_font_color}
\alias{get_optimal_font_color}
\title{Compute optimal font color (black or white) for given hex background colors}
\usage{
get_optimal_font_color(hex_colors)
}
\arguments{
\item{hex_colors}{A character vector of hex color codes (with or without # prefix)}
}
\value{
A character vector of "black" or "white" indicating the optimal font color for each input
}
\description{
This function determines whether black or white text would be most readable
on a given background color by calculating the relative luminance of the background color.
}
\details{
Get optimal font color for hex colors
}
