% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_objects.R
\name{get_question_dimensions}
\alias{get_question_dimensions}
\title{Get default dimensions for question types}
\usage{
get_question_dimensions(tipus)
}
\arguments{
\item{tipus}{A string representing the question type}
}
\value{
A list with width and height values
}
\description{
Get default dimensions for question types
}
