% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_pres_render.r
\name{get_question_plots}
\alias{get_question_plots}
\title{Generate plot chunks for all questions}
\usage{
get_question_plots(
  Survey,
  question_indices = NULL,
  fig_width = 8,
  fig_height = 4
)
}
\arguments{
\item{Survey}{Survey object}

\item{question_indices}{Which questions to render}

\item{fig_width}{Figure width}

\item{fig_height}{Figure height}
}
\value{
List of plot objects
}
\description{
Generate plot chunks for all questions
}
