% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_time_values.R
\name{get_time_values}
\alias{get_time_values}
\title{Get filtered time values}
\usage{
get_time_values(
  df,
  time_type_out,
  time_type_col = "Időszak bontás",
  time_value_col = "Időszak érték",
  num_days = 180,
  from_start = TRUE
)
}
\arguments{
\item{df}{Dataframe. Dataframe to mutate and convert its date texts to date formats.}

\item{time_type_out}{Character string. Time type to filter, can be Éves, Havi, QTR or Mozgó negyedév..}

\item{time_type_col}{Character string. Colname in df, which corresponds to the time type column.}

\item{time_value_col}{Character string. Colname in df, which corresponds to the time value column.}

\item{num_days}{Number. For extracting the first and last n values that contain not analysable values, a variable of number of days should be applied.}

\item{from_start}{Boolean. If the filtering should be from the start.}
}
\value{
A Date value
}
\description{
\code{get_time_values()} creates a date format from dates as text, based on the type of date e.g. converts 2023M01 to date 2023-01-01
}
