% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_integer_breaks.r
\name{integer_breaks}
\alias{integer_breaks}
\title{Returns a function that computes axis breaks at integer values using \code{pretty} and \code{floor}.}
\usage{
integer_breaks(n = 5, ...)
}
\arguments{
\item{n}{Number of breaks (default is 5).}

\item{...}{Additional arguments passed to \code{pretty}.}
}
\value{
A function that takes a numeric vector and returns integer axis breaks.
}
\description{
Returns a function that computes axis breaks at integer values using \code{pretty} and \code{floor}.
}
