% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_palette.r
\name{long_palette}
\alias{long_palette}
\title{Custom color palette for survey plots}
\usage{
long_palette(n = NULL)
}
\arguments{
\item{n}{Number of colors to return. If NULL, returns all colors.}
}
\value{
A character vector of colors.
}
\description{
Returns a vector of colors for use in survey visualizations. If \code{n} is NULL, returns the whole palette. If \code{n} is a number, returns the first \code{n} colors.
}
