% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_oracle_connect.R
\name{oracle_connect}
\alias{oracle_connect}
\title{Connect to RF Oracle schemes}
\usage{
oracle_connect(schema, setLocale = T)
}
\arguments{
\item{schema}{Character string. Existing scheme in the RF Oracle system}

\item{setLocale}{Boolean. If sys.setenv should be run tto set adequate Hungarian language options.}
}
\value{
An ROracle connection OraConnection object.
}
\description{
\code{oracle_connect()} connects to RF schemes, with host, scheme name and password.
}
