% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_oracle_query_utils.R
\name{oracle_list_tables}
\alias{oracle_list_tables}
\title{List available tables}
\usage{
oracle_list_tables(schema, connection = NULL, subset = NULL)
}
\arguments{
\item{schema}{Character string, optional. Schema name to list tables from.
If NULL, lists tables from the current schema.}

\item{connection}{A valid database connection object}

\item{subset}{A regex expression to filter for relevant table names}
}
\value{
A character vector of table names.
}
\description{
\code{oracle_list_tables()} returns a list of all tables available in the current
schema or a specified schema.
}
\details{
List all tables in the current schema
}
