% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_oracle_query_utils.R
\name{oracle_table_exists}
\alias{oracle_table_exists}
\title{Check table existence}
\usage{
oracle_table_exists(schema, table_name, connection = NULL)
}
\arguments{
\item{schema}{Character string, optional. Schema name to check in.}

\item{table_name}{Character string. Name of the table to check.}

\item{connection}{A valid database connection object}
}
\value{
Logical. TRUE if table exists, FALSE otherwise.
}
\description{
\code{oracle_table_exists()} checks if a specified table exists in the current
schema or a specified schema.
}
\details{
Check if a table exists in the schema
}
