% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_time_value.R
\name{parse_time_value}
\alias{parse_time_value}
\title{Convert date texts to date format}
\usage{
parse_time_value(time_type, time_value)
}
\arguments{
\item{time_type}{Character string. Type can be Éves, QTR, Havi, Mozgó negyedév or *Teljes időszak}

\item{time_value}{Character string. Date as a text format}
}
\value{
A Date value
}
\description{
\code{parse_time_value()} creates a date format from dates as text, based on the type of date e.g. converts 2023M01 to date 2023-01-01
}
