% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_objects.R
\name{plot.SurveyQuestion}
\alias{plot.SurveyQuestion}
\title{Create ggplot for SurveyQuestion object}
\usage{
\method{plot}{SurveyQuestion}(x, type = "ggplot", ...)
}
\arguments{
\item{x}{A SurveyQuestion object}

\item{...}{Additional arguments (unused)}
}
\value{
A ggplot object
}
\description{
Create ggplot for SurveyQuestion object
}
