% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xlsb_custom.R
\name{read_xlsb}
\alias{read_xlsb}
\title{Reading xlsb files}
\usage{
read_xlsb(path, sheet)
}
\arguments{
\item{path}{String The relative or absolute path of the file. Finishes in ".xlsb"}

\item{sheet}{String. Sheet in the xlsb document}
}
\value{
A tibble
}
\description{
\code{read_xlsb()} Reads xlsb files with the help of odbc connections
}
