% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regio_shape}
\alias{regio_shape}
\title{Regional Shape Data}
\format{
An sf object with regional shape data for Hungary:
\describe{
\item{NUTS_NAME}{Region name}
\item{geometry}{Spatial geometry data}
...
}
}
\source{
Hungarian spatial data source
}
\usage{
regio_shape
}
\description{
Spatial data object containing regional boundaries and shape information
for Hungarian regions. Used for creating regional distribution maps
in survey visualization functions.
}
\keyword{datasets}
