% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_pres_render.r
\name{render_all_slides}
\alias{render_all_slides}
\title{Render all survey slides with individual plot dimensions}
\usage{
render_all_slides(
  Survey,
  question_indices = NULL,
  default_fig_width = 17,
  default_fig_height = 10,
  include_title = TRUE
)
}
\arguments{
\item{Survey}{Survey object}

\item{question_indices}{Which questions to render (default: all questions)}

\item{default_fig_width}{Default figure width if question doesn't specify}

\item{default_fig_height}{Default figure height if question doesn't specify}

\item{include_title}{Whether to include the survey title as first slide (default: TRUE)}
}
\description{
Render all survey slides with individual plot dimensions
}
