% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflib-package.R
\docType{package}
\name{rflib-package}
\alias{rflib}
\alias{rflib-package}
\title{rflib: RF utilities}
\description{
Package contains functions and utils for internal use.
}
\author{
\strong{Maintainer}: Balint Mazzag \email{mazzag.balint@researchflow.hu}

}
\keyword{internal}
