% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_objects.R
\name{survey_add_definition}
\alias{survey_add_definition}
\title{Add definition Excel to Survey object and assign tipus for each question}
\usage{
survey_add_definition(
  survey_obj,
  definition_path,
  rewrangle = TRUE,
  replot = FALSE
)
}
\arguments{
\item{survey_obj}{A Survey object}

\item{definition_path}{Path to Excel containing columns: kerdes_szam, tipus}

\item{rewrangle}{Logical, rerun wrangling immediately (default: TRUE)}

\item{replot}{Logical, rerun plot creation if enabled (default: FALSE)}
}
\value{
Updated Survey object
}
\description{
Add definition Excel to Survey object and assign tipus for each question
}
