% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_objects.R
\name{survey_add_question}
\alias{survey_add_question}
\title{Add question to Survey object}
\usage{
survey_add_question(survey_obj, question_obj)
}
\arguments{
\item{survey_obj}{A Survey object}

\item{question_obj}{A SurveyQuestion object}
}
\value{
Updated Survey object
}
\description{
Add question to Survey object
}
