% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_echarts_templates.R
\name{survey_echarts_col_eloszlas_total}
\alias{survey_echarts_col_eloszlas_total}
\title{ECharts template for column distribution (total) survey question}
\usage{
survey_echarts_col_eloszlas_total(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
An echarts4r object
}
\description{
Creates a bar chart for column distribution (total) survey questions using echarts4r.
}
