% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_echarts_templates.R
\name{survey_echarts_likert_scale}
\alias{survey_echarts_likert_scale}
\title{ECharts template for Likert scale survey question}
\usage{
survey_echarts_likert_scale(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
An echarts4r object
}
\description{
Creates a bar and scatter chart for Likert scale survey questions using echarts4r.
}
