% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_echarts_templates.R
\name{survey_echarts_likert_scale_rev}
\alias{survey_echarts_likert_scale_rev}
\title{ECharts template for reversed Likert scale survey question}
\usage{
survey_echarts_likert_scale_rev(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
An echarts4r object
}
\description{
Creates a plot for reversed Likert scale survey questions using echarts4r.
}
