% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_echarts_templates.R
\name{survey_echarts_likert_scale_table}
\alias{survey_echarts_likert_scale_table}
\title{ECharts template for Likert scale table survey question}
\usage{
survey_echarts_likert_scale_table(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
An echarts4r object
}
\description{
Creates a plot for Likert scale table survey questions using echarts4r.
}
