% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_echarts_templates.R
\name{survey_echarts_resz_egesz_total}
\alias{survey_echarts_resz_egesz_total}
\title{ECharts template for total part-whole survey question}
\usage{
survey_echarts_resz_egesz_total(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
An echarts4r object
}
\description{
Creates a bar chart for total part-whole survey questions using echarts4r.
}
