% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_echarts_templates.R
\name{survey_echarts_szoveg_buborek_multiple}
\alias{survey_echarts_szoveg_buborek_multiple}
\title{ECharts template for text bubble multiple survey question}
\usage{
survey_echarts_szoveg_buborek_multiple(question, language = "hu")
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}

\item{language}{Language code for chart labels and formatting (default: "hu" for Hungarian)}
}
\value{
An echarts4r object
}
\description{
Creates a word cloud for text bubble multiple survey questions using echarts4r.
}
