% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_echarts_templates.R
\name{survey_echarts_szoveg_col_egyeb_new}
\alias{survey_echarts_szoveg_col_egyeb_new}
\title{ECharts template for new other text column survey question}
\usage{
survey_echarts_szoveg_col_egyeb_new(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
An echarts4r object
}
\description{
Creates a plot for new other text column survey questions using echarts4r.
}
