% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_echarts_templates.R
\name{survey_echarts_table_egyeb}
\alias{survey_echarts_table_egyeb}
\title{ECharts template for other table survey question}
\usage{
survey_echarts_table_egyeb(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
An echarts4r object
}
\description{
Creates a plot for other table survey questions using echarts4r.
}
