% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_echarts_templates.R
\name{survey_echarts_table_oszlop}
\alias{survey_echarts_table_oszlop}
\title{ECharts template for table column survey question}
\usage{
survey_echarts_table_oszlop(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
An echarts4r object
}
\description{
Creates a plot for table column survey questions using echarts4r.
}
