% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_echarts_templates.R
\name{survey_echarts_year_eloszlas_unscale}
\alias{survey_echarts_year_eloszlas_unscale}
\title{ECharts template for year distribution (unscaled) survey question}
\usage{
survey_echarts_year_eloszlas_unscale(question, language = "hu")
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}

\item{language}{Language code for chart labels and formatting (default: "hu" for Hungarian)}
}
\value{
An echarts4r object
}
\description{
Creates a histogram for year distribution (unscaled) survey questions using echarts4r.
}
