% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_ggplot_templates.R
\name{survey_ggplot_likert_scale}
\alias{survey_ggplot_likert_scale}
\title{GGPlot template for Likert scale survey question}
\usage{
survey_ggplot_likert_scale(question, language = "hu")
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}

\item{language}{Language code for chart labels and formatting (default: "hu" for Hungarian)}
}
\value{
A ggplot object
}
\description{
GGPlot template for Likert scale survey question
}
