% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_ggplot_templates.R
\name{survey_ggplot_likert_scale_ref}
\alias{survey_ggplot_likert_scale_ref}
\title{GGPlot template for Likert scale reference survey question}
\usage{
survey_ggplot_likert_scale_ref(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
A ggplot object
}
\description{
GGPlot template for Likert scale reference survey question
}
