% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_ggplot_templates.R
\name{survey_ggplot_naptar}
\alias{survey_ggplot_naptar}
\title{GGPlot template for multiple part-whole survey question}
\usage{
survey_ggplot_naptar(question, language = "hu")
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}

\item{language}{Language code for chart labels and formatting (default: "hu" for Hungarian)}
}
\value{
A ggplot or grid object
}
\description{
GGPlot template for multiple part-whole survey question
}
