% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_ggplot_templates.R
\name{survey_ggplot_resz_egesz_multiple}
\alias{survey_ggplot_resz_egesz_multiple}
\title{GGPlot template for multiple part-whole survey question}
\usage{
survey_ggplot_resz_egesz_multiple(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
A ggplot or grid object
}
\description{
GGPlot template for multiple part-whole survey question
}
