% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_ggplot_templates.R
\name{survey_ggplot_resz_egesz_total}
\alias{survey_ggplot_resz_egesz_total}
\title{GGPlot template for total part-whole survey question}
\usage{
survey_ggplot_resz_egesz_total(question, language = "hu")
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}

\item{language}{Language code for chart labels and formatting (default: "hu" for Hungarian)}
}
\value{
A ggplot or grid object
}
\description{
GGPlot template for total part-whole survey question
}
