% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_ggplot_templates.R
\name{survey_ggplot_szoveg_buborek}
\alias{survey_ggplot_szoveg_buborek}
\title{GGPlot template for text bubble multiple survey question}
\usage{
survey_ggplot_szoveg_buborek(question, language = "hu")

survey_ggplot_szoveg_buborek(question, language = "hu")
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}

\item{language}{Language code for chart labels and formatting (default: "hu" for Hungarian)}
}
\value{
A ggplot object

A ggplot object
}
\description{
GGPlot template for text bubble multiple survey question

GGPlot template for text bubble multiple survey question
}
