% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_ggplot_templates.R
\name{survey_ggplot_szoveg_col_egyeb_new}
\alias{survey_ggplot_szoveg_col_egyeb_new}
\title{GGPlot template for new other text column survey question}
\usage{
survey_ggplot_szoveg_col_egyeb_new(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
A ggplot object
}
\description{
GGPlot template for new other text column survey question
}
