% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_ggplot_templates.R
\name{survey_ggplot_year_eloszlas}
\alias{survey_ggplot_year_eloszlas}
\title{GGPlot template for year distribution survey question}
\usage{
survey_ggplot_year_eloszlas(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
A ggplot object
}
\description{
GGPlot template for year distribution survey question
}
