% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_init.R
\name{survey_init}
\alias{survey_init}
\title{Initialize survey workflow as Survey object}
\usage{
survey_init(path, clean_text = TRUE, save_output = FALSE)
}
\arguments{
\item{path}{Path to the Excel file}

\item{clean_text}{Whether to run text classification on open-text (default = TRUE)}

\item{save_output}{Whether to save intermediate outputs (default = FALSE)}
}
\value{
A Survey object with metadata, labels, and question objects
}
\description{
Initialize survey workflow as Survey object
}
