% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_load_excel.R
\name{survey_load_title}
\alias{survey_load_title}
\title{Load title string from the Title sheet (cell A2) in the survey Excel}
\usage{
survey_load_title(path)
}
\arguments{
\item{path}{Path to the Excel file}
}
\value{
A character string with the title from cell A2
}
\description{
Load title string from the Title sheet (cell A2) in the survey Excel
}
