% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_wrangle_functions.r
\name{survey_plot_dispatch}
\alias{survey_plot_dispatch}
\title{Dispatch plot function based on plot type and question type}
\usage{
survey_plot_dispatch(plottype, question)
}
\arguments{
\item{plottype}{A string: "ggplot" or "echarts"}

\item{question}{A SurveyQuestion object}
}
\value{
A plot object (ggplot or echarts4r)
}
\description{
Dispatch plot function based on plot type and question type
}
