% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_pair_questions_labels_advanced.R
\name{survey_prepare_answers}
\alias{survey_prepare_answers}
\title{Create a tibble of original and cleaned names}
\usage{
survey_prepare_answers(df_answers)
}
\arguments{
\item{df_answers}{Character vector of original names}
}
\value{
A tibble with columns 'name' and 'name_clean'
}
\description{
Create a tibble of original and cleaned names
}
