% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_pair_questions_labels_advanced.R
\name{survey_prepare_label}
\alias{survey_prepare_label}
\title{Prepare label table for advanced pairing}
\usage{
survey_prepare_label(df_labels)
}
\arguments{
\item{df_labels}{Data frame of label definitions}
}
\value{
A tibble with cleaned and extracted label info
}
\description{
Prepare label table for advanced pairing
}
