% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_wrangle_functions.r
\name{survey_wrangle_col_eloszlas_total}
\alias{survey_wrangle_col_eloszlas_total}
\title{Wrangle column distribution question}
\usage{
survey_wrangle_col_eloszlas_total(df, labels)
}
\arguments{
\item{df}{A tibble with columns respondent_id and answer}

\item{labels}{A tibble with question labels}
}
\value{
A tibble with frequency per answer
}
\description{
Wrangle column distribution question
}
