% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_wrangle_functions.r
\name{survey_wrangle_combobox}
\alias{survey_wrangle_combobox}
\title{Wrangle combobox question}
\usage{
survey_wrangle_combobox(df, labels)
}
\arguments{
\item{df}{A tibble with columns respondent_id, answer, kerdes}

\item{labels}{A tibble with question labels}
}
\value{
Not implemented
}
\description{
Wrangle combobox question
}
