% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_wrangle_functions.r
\name{survey_wrangle_dispatch}
\alias{survey_wrangle_dispatch}
\title{Dispatch wrangling function based on question type}
\usage{
survey_wrangle_dispatch(tipus, data, labels)
}
\arguments{
\item{tipus}{A string defining the question type}

\item{data}{A dataframe with the question's data}

\item{labels}{A dataframe with the question's labels}
}
\value{
A wrangled dataframe suitable for plotting or analysis
}
\description{
Dispatch wrangling function based on question type
}
