% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_wrangle_functions.r
\name{survey_wrangle_diszkret_eloszlas}
\alias{survey_wrangle_diszkret_eloszlas}
\title{Wrangle discrete distribution question}
\usage{
survey_wrangle_diszkret_eloszlas(df, labels)
}
\arguments{
\item{df}{A tibble with columns respondent_id, answer, kerdes}

\item{labels}{A tibble with question labels}
}
\value{
Not implemented
}
\description{
Wrangle discrete distribution question
}
