% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_wrangle_functions.r
\name{survey_wrangle_likert_scale}
\alias{survey_wrangle_likert_scale}
\title{Wrangle Likert scale question}
\usage{
survey_wrangle_likert_scale(df, labels)
}
\arguments{
\item{df}{A tibble with columns respondent_id and answer}

\item{labels}{A tibble with question labels}
}
\value{
A tibble with summarized proportions per answer option
}
\description{
Wrangle Likert scale question
}
