% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_wrangle_functions.r
\name{survey_wrangle_likert_scale_table}
\alias{survey_wrangle_likert_scale_table}
\title{Wrangle Likert scale table question}
\usage{
survey_wrangle_likert_scale_table(df, labels)
}
\arguments{
\item{df}{A tibble with columns respondent_id, answer, kerdes}

\item{labels}{A tibble with question labels}
}
\value{
Not implemented
}
\description{
Wrangle Likert scale table question
}
