% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_wrangle_functions.r
\name{survey_wrangle_resz_egesz_multiple}
\alias{survey_wrangle_resz_egesz_multiple}
\title{Wrangle multiple part-whole question}
\usage{
survey_wrangle_resz_egesz_multiple(df, labels)
}
\arguments{
\item{df}{A tibble with columns respondent_id and answer}

\item{labels}{A tibble with question labels}
}
\value{
A tibble with proportions for each selected part
}
\description{
Wrangle multiple part-whole question
}
