% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_wrangle_functions.r
\name{survey_wrangle_resz_egesz_total}
\alias{survey_wrangle_resz_egesz_total}
\title{Wrangle total part-whole question}
\usage{
survey_wrangle_resz_egesz_total(df, labels)
}
\arguments{
\item{df}{A tibble with columns respondent_id and answer}

\item{labels}{A tibble with question labels}
}
\value{
A tibble with one row per respondent and numeric total
}
\description{
Wrangle total part-whole question
}
