% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_wrangle_functions.r
\name{survey_wrangle_szam_col_egyeb}
\alias{survey_wrangle_szam_col_egyeb}
\title{Wrangle other numeric column question}
\usage{
survey_wrangle_szam_col_egyeb(df, labels)
}
\arguments{
\item{df}{A tibble with columns respondent_id, answer, kerdes}

\item{labels}{A tibble with question labels}
}
\value{
Not implemented
}
\description{
Wrangle other numeric column question
}
