% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_wrangle_functions.r
\name{survey_wrangle_szoveg_buborek_multiple}
\alias{survey_wrangle_szoveg_buborek_multiple}
\title{Wrangle text bubble multiple question}
\usage{
survey_wrangle_szoveg_buborek_multiple(df, labels)
}
\arguments{
\item{df}{A tibble with columns respondent_id, answer, kerdes}

\item{labels}{A tibble with question labels}
}
\value{
A tibble with processed text bubble multiple answers
}
\description{
Wrangle text bubble multiple question
}
