% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_wrangle_functions.r
\name{survey_wrangle_szoveg_col_egyeb_new}
\alias{survey_wrangle_szoveg_col_egyeb_new}
\title{Wrangle new other text column question}
\usage{
survey_wrangle_szoveg_col_egyeb_new(df, labels)
}
\arguments{
\item{df}{A tibble with columns respondent_id, answer, kerdes}

\item{labels}{A tibble with question labels}
}
\value{
Not implemented
}
\description{
Wrangle new other text column question
}
