% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_wrangle_functions.r
\name{survey_wrangle_table_oszlop}
\alias{survey_wrangle_table_oszlop}
\title{Wrangle table column question}
\usage{
survey_wrangle_table_oszlop(df, labels)
}
\arguments{
\item{df}{A tibble with columns respondent_id, answer, kerdes}

\item{labels}{A tibble with question labels}
}
\value{
Not implemented
}
\description{
Wrangle table column question
}
