% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_wrangle_functions.r
\name{survey_wrangle_terulet_eloszlas}
\alias{survey_wrangle_terulet_eloszlas}
\title{Wrangle area distribution question}
\usage{
survey_wrangle_terulet_eloszlas(df, labels)
}
\arguments{
\item{df}{A tibble with columns respondent_id, answer, kerdes}

\item{labels}{A tibble with question labels}
}
\value{
Not implemented
}
\description{
Wrangle area distribution question
}
