% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_wrangle_functions.r
\name{survey_wrangle_year_eloszlas_unscale}
\alias{survey_wrangle_year_eloszlas_unscale}
\title{Wrangle year distribution (unscaled)}
\usage{
survey_wrangle_year_eloszlas_unscale(df, labels)
}
\arguments{
\item{df}{A tibble with columns respondent_id and answer}

\item{labels}{A tibble with question labels}
}
\value{
A tibble with numeric answers joined to labels
}
\description{
Wrangle year distribution (unscaled)
}
