% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teruleti_szotar}
\alias{teruleti_szotar}
\title{Territorial Dictionary}
\format{
A data frame with territorial mapping information:
\describe{
\item{code}{Territorial code}
\item{name}{Territorial name}
\item{region}{Regional classification}
...
}
}
\source{
Hungarian territorial classification system
}
\usage{
teruleti_szotar
}
\description{
Dictionary data containing territorial codes and names for Hungarian
administrative regions. Used for mapping territorial data in survey
analysis and visualization functions.
}
\keyword{datasets}
