% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloszlasPlots.R
\name{validate_columns}
\alias{validate_columns}
\title{Validate Required Columns}
\usage{
validate_columns(data, required_columns)
}
\arguments{
\item{data}{A data frame to validate.}

\item{required_columns}{A vector of column names that must be present in the data frame.}
}
\value{
None. Stops with an error if required columns are missing.
}
\description{
Validate Required Columns
}
