% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_time_periods.R
\name{validate_time_periods}
\alias{validate_time_periods}
\title{Validate time periods in tables}
\usage{
validate_time_periods(idoszak_ertek, idoszak_bontas, min_date, max_date)
}
\arguments{
\item{idoszak_ertek}{Character vector. Tiem period values to analyse}

\item{idoszak_bontas}{Character vector with values as: "Éves", "Havi", "QTR", "*Teljes időszak"}

\item{min_date}{Date string (YYYY.MM.DD.), will be converted into date. Gives the minimum date for the search.}

\item{max_date}{Date string (YYYY.MM.DD.), will be converted into date. Gives the maximum date for the search.}
}
\value{
A vector with correct labels and NA for out of date range values.
}
\description{
\code{validate_time_periods()} gets latest table lists from Oracle connections.
}
