% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_pres_generate.r
\name{batch_generate_presentations}
\alias{batch_generate_presentations}
\title{Batch generate presentations for multiple surveys}
\usage{
batch_generate_presentations(
  survey_folders,
  titles = NULL,
  base_subtitle = "Vizualizációk"
)
}
\arguments{
\item{survey_folders}{Vector of survey folder names}

\item{titles}{Vector of survey titles}

\item{base_subtitle}{Base subtitle to use}
}
\value{
Vector of generated file paths
}
\description{
Batch generate presentations for multiple surveys
}
