% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_table.R
\name{get_latest_table}
\alias{get_latest_table}
\title{Get latest tables from an Oracle scheme}
\usage{
get_latest_table(
  table_name_val,
  connection_val,
  tibble_of_tables = NULL,
  list_tables = F
)
}
\arguments{
\item{table_name_val}{Character string. Filtering string to table names}

\item{connection_val}{List of OraConnection objects. In which we search for table names.}

\item{tibble_of_tables}{Tibble. If previously all possible names are extracted, they can be set as default.}

\item{list_tables}{Boolean. If return should be total table names or only the latest ones.}
}
\value{
A tibble with given table name patterns and their latest table names in an Oracle scheme.
}
\description{
\code{get_latest_table()} gets latest table lists from Oracle connections.
}
