% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_pres_render.r
\name{render_plot_with_dimensions}
\alias{render_plot_with_dimensions}
\title{Render a plot with specific dimensions using permanent files}
\usage{
render_plot_with_dimensions(id, width, height)
}
\arguments{
\item{width}{Figure width in inches}

\item{height}{Figure height in inches}

\item{plot}{ggplot object}

\item{label}{Unique label for the plot}
}
\description{
Render a plot with specific dimensions using permanent files
}
