% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_echarts_templates.R
\name{survey_echarts_col_eloszlas}
\alias{survey_echarts_col_eloszlas}
\title{ECharts template for column distribution survey question}
\usage{
survey_echarts_col_eloszlas(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
An echarts4r object
}
\description{
Creates a plot for column distribution survey questions using echarts4r.
}
