% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_echarts_templates.R
\name{survey_echarts_diszkret_eloszlas}
\alias{survey_echarts_diszkret_eloszlas}
\title{ECharts template for discrete distribution survey question}
\usage{
survey_echarts_diszkret_eloszlas(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
An echarts4r object
}
\description{
Creates a plot for discrete distribution survey questions using echarts4r.
}
