% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_echarts_templates.R
\name{survey_echarts_hoterkep}
\alias{survey_echarts_hoterkep}
\title{ECharts template for heatmap survey question}
\usage{
survey_echarts_hoterkep(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
An echarts4r object
}
\description{
Creates a heatmap for survey questions using echarts4r.
}
