% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_echarts_templates.R
\name{survey_echarts_likert_scale_ref}
\alias{survey_echarts_likert_scale_ref}
\title{ECharts template for Likert scale reference survey question}
\usage{
survey_echarts_likert_scale_ref(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
An echarts4r object
}
\description{
Creates a plot for Likert scale reference survey questions using echarts4r.
}
