% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_echarts_templates.R
\name{survey_echarts_table}
\alias{survey_echarts_table}
\title{ECharts template for table survey question}
\usage{
survey_echarts_table(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
An echarts4r object
}
\description{
Creates a heatmap for table survey questions using echarts4r.
}
