% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_ggplot_templates.R
\name{survey_ggplot_col_eloszlas_multiple}
\alias{survey_ggplot_col_eloszlas_multiple}
\title{GGPlot template for column distribution (multiple) survey question}
\usage{
survey_ggplot_col_eloszlas_multiple(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
A ggplot object
}
\description{
GGPlot template for column distribution (multiple) survey question
}
