% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_ggplot_templates.R
\name{survey_ggplot_resz_egesz_total}
\alias{survey_ggplot_resz_egesz_total}
\title{GGPlot template for total part-whole survey question}
\usage{
survey_ggplot_resz_egesz_total(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
A ggplot or grid object
}
\description{
GGPlot template for total part-whole survey question
}
