% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_ggplot_templates.R
\name{survey_ggplot_szoveg_buborek_multiple}
\alias{survey_ggplot_szoveg_buborek_multiple}
\title{GGPlot template for text bubble multiple survey question}
\usage{
survey_ggplot_szoveg_buborek_multiple(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
A ggplot object
}
\description{
GGPlot template for text bubble multiple survey question
}
