% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_ggplot_templates.R
\name{survey_ggplot_table_rev}
\alias{survey_ggplot_table_rev}
\title{GGPlot template for table reversed survey question}
\usage{
survey_ggplot_table_rev(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
A ggplot object
}
\description{
GGPlot template for table reversed survey question
}
