% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_ggplot_templates.R
\name{survey_ggplot_year_eloszlas_unscale}
\alias{survey_ggplot_year_eloszlas_unscale}
\title{GGPlot template for year distribution (unscaled) survey question}
\usage{
survey_ggplot_year_eloszlas_unscale(question)
}
\arguments{
\item{question}{A list containing wrangled data and plot parameters}
}
\value{
A ggplot object
}
\description{
GGPlot template for year distribution (unscaled) survey question
}
