% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teruleti_szotar}
\alias{teruleti_szotar}
\title{Regional Dictionary Dataset}
\format{
A data frame with 3 variables:
\describe{
\item{postal_code}{Character. Hungarian postal codes}
\item{county}{Character. County names}
\item{region}{Character. Region names}
}
}
\source{
Internal research data
}
\usage{
teruleti_szotar
}
\description{
A dataset containing the mapping between postal codes, counties, and regions in Hungary.
}
\keyword{datasets}
