% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teruleti_szotar}
\alias{teruleti_szotar}
\title{Regional Dictionary Dataset}
\format{
A data frame with 3 variables:
\describe{
\item{iranyitoszam}{Character. Hungarian postal codes}
\item{megye_nev}{Character. County names}
\item{regio_nev}{Character. Region names}
\item{telepules_nev}{Character. Settlement names}
}
}
\source{
Internal research data
}
\usage{
teruleti_szotar
}
\description{
A dataset containing the mapping between postal codes, counties, and regions in Hungary.
}
\keyword{datasets}
